confirm delete alert

+++ detail.php
@@ -58,19 +58,19 @@ $post = $posts[0];
                 <div class="keterangan">
                     <p>
                     <p style="align-self: cente">
-                    <p>Posted by: <a href="profil.php?idUser=<?= $post['id_user'] ?>"><?= $post['username'] ?></a></p>
+                    <p>Diposting oleh: <strong><a href="profil.php?idUser=<?= $post['id_user'] ?>"><?= $post['username'] ?></a></strong></p>
                     <!-- like -->
 
                     <!-- donlot -->
                     <a id="downloadBtn" href="<?= $post['image'] ?>" download class="download-btn">Unduh Gambar</a>
                     <!-- dlete dan edit -->
-                     <div class="edit-hapus">
-                         <?php if (isset($idUser) && $idUser == $post['id_user']): ?>
-                             <a href="editPost.php?id_post=<?= $post['id_post'] ?>" class="edit-btn">Edit</a>
-                             <a href="deletePost.php?id_post=<?= $post['id_post'] ?>" class="delete-btn">Hapus</a>
-                         <?php endif; ?>
-                         <p style="font-size:32px"><i class='bx bx-comment'></i> <?= $post['comment_count'] ?></p>
-                     </div>
+                    <div class="edit-hapus">
+                        <?php if (isset($idUser) && $idUser == $post['id_user']): ?>
+                            <a href="editPost.php?id_post=<?= $post['id_post'] ?>" class="edit-btn">Edit</a>
+                            <a href="deletePost.php?id_post=<?= $post['id_post'] ?>" class="delete-btn" onclick="return confirm('Apakah Anda yakin ingin menghapus postingan ini?');">Hapus</a>
+                        <?php endif; ?>
+                        <p style="font-size:32px"><i class='bx bx-comment'></i> <?= $post['comment_count'] ?></p>
+                    </div>
                 </div>
 
                 <!-- komenn -->
